pyproject.toml
[project]
name = "crash-detection-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=1.8.3",
    "numpy==2.4.1",
    "opencv-python-headless>=4.13.0.90",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-box>=7.3.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "typeguard>=4.4.4",
]
